Moves api response drawing to it's new home.
